Credit "the paper" from which the PCLMULQDQ specialization came from
While "Fast CRC computation for generic polynomials using PCLMULQDQ
instruction" is a well-known paper in the realm of specialized CRC and
compression algorithms for the x86 architecture that has seen widespread
implementation, not everyone is aware of it, and it's nevertheless a
good practice to cite the sources ideas and knowledge come from.

Let's add a comment to the `pclmulqdq` module citing the whitepaper,
including a link to a mirror hosting its PDF, for easy access by
prospective contributors and other interested parties.